Moving various utils into single utils file. Renaming fuzz files
