+++ src/agent0/ethpy/base/__init__.py
@@ -3,7 +3,6 @@
 from .abi import load_abi_from_file, load_all_abis
 from .errors import ABIError, UnknownBlockError, decode_error_selector_for_contract
 from .receipts import get_event_object, get_transaction_logs
-from .retry_utils import retry_call
 from .rpc_interface import get_account_balance, set_anvil_account_balance
 from .transactions import (
     async_eth_transfer,
