+++ src/agent0/hyperfuzz/system_fuzz/__init__.py
@@ -1,4 +1,4 @@
 """System level fuzz bots."""
 
 from .invariant_checks import run_invariant_checks
-from .run_local_fuzz_bots import generate_fuzz_hyperdrive_config, run_local_fuzz_bots
+from .run_fuzz_bots import generate_fuzz_hyperdrive_config, run_fuzz_bots
