+++ src/agent0/utils_test.py
@@ -1,9 +1,30 @@
 """Test executing transactions."""
 
+import asyncio
+from functools import partial
+
 import pytest
 
 from agent0.core.hyperdrive.interactive import LocalChain
-from agent0.utils import retry_call
+
+from .utils import async_runner, retry_call
+
+
+def test_async_runner():
+    def _return_int(i: int, j: int) -> tuple[int, int]:
+        return (i, j)
+
+    # TODO because _async_runner only takes one set of arguments for all calls,
+    # we make partial calls for each call. The proper fix here is to generalize
+    # _async_runner to take separate arguments for each call.
+    partials = [partial(_return_int, i) for i in [1, 2, 3]]
+
+    out_tuples = asyncio.run(async_runner(return_exceptions=False, funcs=partials, j=10))
+
+    for i, o in enumerate(out_tuples):
+        assert len(o) == 2
+        assert o[0] == i
+        assert o[1] == 10
 
 
 @pytest.mark.anvil
